refactor(suite-native): split portfolio line graph
